% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{FAST.R}
\alias{FAST.R}
\title{FAST.R: Analyze and Visualize FAST-Generated Data}
\usage{
FAST.R()
}
\value{
Launches an R shiny app to analyze and visualize data generated with the FAST workflow
}
\description{
R shiny app to perform data analysis and visualization for the
Fully Automated Senescence Test (FAST) workflow.
}
\examples{
if(interactive()){FAST.R()}
}
