% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{FASTR}
\alias{FASTR}
\title{FASTR: Analyze and Visualize FAST-Generated Data}
\usage{
FASTR()
}
\value{
Launches an R shiny app to analyze and visualize data generated with the FAST workflow
}
\description{
R shiny app to perform data analysis and visualization for the
Fully Automated Senescence Test (FAST) workflow.
}
\examples{
\dontrun{FASTR()}
}
